requirements/ci.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    ./bin/compile_dependencies.sh
#
asgiref==3.5.2
    # via
    #   -r requirements/base.txt
    #   django
astroid==1.6.6
    # via pylint
async-timeout==4.0.2
    # via
    #   -r requirements/base.txt
    #   redis
attrs==22.1.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
black==22.6.0
    # via -r requirements/testing.in
bleach==5.0.1
    # via
    #   -r requirements/base.txt
    #   django-markup
certifi==2022.6.15
    # via
    #   -r requirements/base.txt
    #   requests
charset-normalizer==2.1.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via black
codecov==2.1.12
    # via -r requirements/ci.in
coverage==6.4.4
    # via
    #   -r requirements/ci.in
    #   codecov
deprecated==1.2.13
    # via
    #   -r requirements/base.txt
    #   redis
django==3.2.14
    # via
    #   -r requirements/base.txt
    #   django-admin-index
    #   django-axes
    #   django-choices
    #   django-filter
    #   django-hijack
    #   django-markup
    #   django-redis
    #   django-relativedelta
    #   django-rest-framework-condition
    #   django-rosetta
    #   django-solo
    #   djangorestframework
    #   drf-nested-routers
    #   drf-spectacular
    #   vng-api-common
django-admin-index==2.0.2
    # via -r requirements/base.txt
django-axes==5.35.0
    # via -r requirements/base.txt
django-better-admin-arrayfield==1.4.2
    # via -r requirements/base.txt
django-choices==1.7.2
    # via
    #   -r requirements/base.txt
    #   vng-api-common
django-compat==1.0.15
    # via
    #   -r requirements/base.txt
    #   django-hijack-admin
django-cors-middleware==1.5.0
    # via -r requirements/base.txt
django-filter==21.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
django-hijack==3.2.1
    # via
    #   -r requirements/base.txt
    #   django-hijack-admin
django-hijack-admin==2.1.10
    # via -r requirements/base.txt
django-ipware==4.0.2
    # via
    #   -r requirements/base.txt
    #   django-axes
django-markup[all_filter_dependencies]==1.5
    # via -r requirements/base.txt
django-ordered-model==3.6
    # via
    #   -r requirements/base.txt
    #   django-admin-index
django-redis==5.2.0
    # via -r requirements/base.txt
django-relativedelta==1.1.2
    # via -r requirements/base.txt
django-rest-framework-condition==0.1.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
django-rosetta==0.9.8
    # via -r requirements/base.txt
django-solo==2.0.0
    # via
    #   -r requirements/base.txt
    #   vng-api-common
djangorestframework==3.12.4
    # via
    #   -r requirements/base.txt
    #   drf-nested-routers
    #   drf-spectacular
    #   vng-api-common
djangorestframework-camel-case==1.3.0
    # via
    #   -r requirements/base.txt
    #   vng-api-common
docutils==0.19
    # via
    #   -r requirements/base.txt
    #   django-markup
    #   python-creole
drf-flex-fields==0.9.8
    # via -r requirements/base.txt
drf-nested-routers==0.93.4
    # via
    #   -r requirements/base.txt
    #   vng-api-common
drf-spectacular==0.23.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
drf-writable-nested==0.6.3
    # via -r requirements/base.txt
factory-boy==3.2.1
    # via -r requirements/testing.in
faker==14.1.0
    # via factory-boy
freezegun==1.2.2
    # via -r requirements/testing.in
gemma-zds-client==1.0.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
html5lib==1.1
    # via
    #   -r requirements/base.txt
    #   textile
idna==3.3
    # via
    #   -r requirements/base.txt
    #   requests
importlib-metadata==4.12.0
    # via
    #   -r requirements/base.txt
    #   markdown
inflection==0.5.1
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
iso-639==0.4.5
    # via
    #   -r requirements/base.txt
    #   vng-api-common
isodate==0.6.1
    # via
    #   -r requirements/base.txt
    #   vng-api-common
isort==5.10.1
    # via
    #   -r requirements/testing.in
    #   pylint
jsonschema==4.15.0
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
lazy-object-proxy==1.7.1
    # via astroid
markdown==3.3.7
    # via
    #   -r requirements/base.txt
    #   django-markup
mccabe==0.7.0
    # via pylint
mypy-extensions==0.4.3
    # via black
oyaml==1.0
    # via
    #   -r requirements/base.txt
    #   vng-api-common
packaging==21.3
    # via
    #   -r requirements/base.txt
    #   redis
pathspec==0.9.0
    # via black
pep8==1.7.1
    # via -r requirements/testing.in
platformdirs==2.5.2
    # via black
polib==1.1.1
    # via
    #   -r requirements/base.txt
    #   django-rosetta
psycopg2==2.9.3
    # via -r requirements/base.txt
pygments==2.12.0
    # via
    #   -r requirements/base.txt
    #   django-markup
pyjwt==2.4.0
    # via
    #   -r requirements/base.txt
    #   gemma-zds-client
    #   vng-api-common
pylint==1.9.2
    # via -r requirements/testing.in
pyparsing==3.0.9
    # via
    #   -r requirements/base.txt
    #   packaging
pyrsistent==0.18.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
python-creole==1.4.10
    # via
    #   -r requirements/base.txt
    #   django-markup
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   django-relativedelta
    #   faker
    #   freezegun
python-dotenv==0.20.0
    # via -r requirements/base.txt
pytz==2022.1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==6.0
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
    #   gemma-zds-client
    #   oyaml
    #   vng-api-common
raven==6.10.0
    # via -r requirements/base.txt
redis==4.3.4
    # via
    #   -r requirements/base.txt
    #   django-redis
regex==2022.7.25
    # via
    #   -r requirements/base.txt
    #   textile
requests==2.28.1
    # via
    #   -r requirements/base.txt
    #   codecov
    #   django-rosetta
    #   gemma-zds-client
    #   requests-mock
    #   vng-api-common
requests-mock==1.9.3
    # via -r requirements/testing.in
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   astroid
    #   bleach
    #   django-choices
    #   django-compat
    #   html5lib
    #   isodate
    #   pylint
    #   python-dateutil
    #   requests-mock
smartypants==2.0.1
    # via
    #   -r requirements/base.txt
    #   django-markup
sqlparse==0.4.2
    # via
    #   -r requirements/base.txt
    #   django
tblib==1.7.0
    # via -r requirements/testing.in
textile==4.0.2
    # via
    #   -r requirements/base.txt
    #   django-markup
tomli==2.0.1
    # via black
typing-extensions==4.3.0
    # via black
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
urllib3==1.26.10
    # via
    #   -r requirements/base.txt
    #   requests
vng_api_common @ git+https://github.com//MatthijsBekendam/vng-api-common.git@drf_compatibilty_newer_django_fix
    # via -r requirements/base.txt
webencodings==0.5.1
    # via
    #   -r requirements/base.txt
    #   bleach
    #   html5lib
wrapt==1.14.1
    # via
    #   -r requirements/base.txt
    #   astroid
    #   deprecated
zipp==3.8.0
    # via
    #   -r requirements/base.txt
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==63.1.0
    # via
    #   -r requirements/base.txt
    #   django-axes
